Struct csil_schedule_shared::shift::Shift

pub struct Shift {
    pub id: Uuid,
    pub date: NaiveDate,
    pub start_time: ChicagoTime,
    pub end_time: ChicagoTime,
    pub tutor_id: Uuid,
    pub cleaning: Vec<CleaningLocation>,
    pub requesting_cover: bool,
}

A CSIL shift

This is the representation that the API and frontend use.

Fields
id: Uuiddate: NaiveDatestart_time: ChicagoTimeend_time: ChicagoTimetutor_id: Uuidcleaning: Vec<CleaningLocation>requesting_cover: bool
Implementations

impl Shift

pub fn overlaps(&self, other: &Shift) -> bool

Returns true if this shift overlaps with other

End points are included, so a shift that ends at 11:00 overlaps with a shift that starts at 11:00. This function also checks that the date field matches, but it ignores the tutor_id field.

See also: properly_overlaps

pub fn properly_overlaps(&self, other: &Shift) -> bool

Returns true if this shift overlaps with other

End point are not included, so a shift that ends at 11:00 does not overlap with a shift that starts at 11:00. This function also checks that the date field matches, but it ignores the tutor_id field.

See also: overlaps

pub fn split(&self, t1: ChicagoTime, t2: ChicagoTime) -> Vec<Shift>

Splits a shift in two places

This can produce up to 3 shifts, but if any shift has length 0, it does not get returned.

Panics

Panics if any of the following are true

t1 < self.start_time

t2 <= t1

self.end_time < t2

pub const fn builder(&self) -> ShiftBuilder<'_>

Returns a ShiftBuilder based on this shift

pub const fn increments(&self, step: u32) -> IncrementIter

Returns an iterator over the increments in this shift

Increments are the chunks of length step between this shift’s start and end time. For example, if a shift starts at 10:00 and ends at 11:00 and step is 15, then the shift’s increments are 10:00, 10:15, 10:30 and 10:45.

let shift = Shift {
    id: Uuid::new_v4(),
    date: NaiveDate::from_ymd_opt(2023, 1, 1).unwrap(),
    start_time: ChicagoTime::new(10, 0).unwrap(),
    end_time: ChicagoTime::new(11, 0).unwrap(),
    tutor_id: Uuid::new_v4(),
    cleaning: Vec::new(),
    status: ShiftStatus::Normal,
};

let increments: Vec<_> = shift.increments(15).collect();

assert_eq!(
    increments,
    vec![
        ChicagoTime::new(10, 0).unwrap(),
        ChicagoTime::new(10, 15).unwrap(),
        ChicagoTime::new(10, 30).unwrap(),
        ChicagoTime::new(10, 45).unwrap(),
    ],
);

pub fn duration(&self) -> u32

Returns the duration of this shift in minutes

pub fn is_empty(&self) -> bool

Returns true if the shift is empty